pin grafana version
current versions like 4.6.0 have extra ssl/https checks
our graphite has a self signed ssl cert, but not for the
hostname 'graphitemon' that we give the container, so grafana
would complain.
I tried to use
https://github.com/graphite-project/docker-graphite-statsd instead
(which basically just requires switching the port to 80, and updating
the config files) but i couldn't get any data into it, so gave up on
that.